+++ lib/grape/endpoint.rb
@@ -403,5 +403,13 @@ module Grape
       options[:options_route_enabled] &&
         env[Grape::Http::Headers::REQUEST_METHOD] == Grape::Http::Headers::OPTIONS
     end
+
+    def method_missing(name, *_args)
+      raise NoMethodError.new("undefined method `#{name}' for #{self.class} in `#{route.origin}' endpoint")
+    end
+
+    def respond_to_missing?(method_name, include_private = false)
+      super
+    end
   end
 end